Library management optimization and scability
